complete day 1 / 2 and create aoc runner platform
the runner searches for all classes that implement
the IProblem interface and runs them with the input files
provided by either the [PuzzleFile] or [SampleFile] attributes.

Output is displayed in a terminal-based table with crude runtime
duration tracking.

Future features to include benchmarking, run history, better CLI
args, and more.